+++ src/modules/teams/actions.ts
@@ -22,5 +22,6 @@ export const updateTeamRoster = (data: IUpdateRosterData | IUpdateRosterData[]) 
 export const assignPlayerToRoster = (teamId: ITeam['id'], player: IPlayer) =>
     action(types.ASSIGN_PLAYER_TO_ROSTER, { teamId, player });
 export const removeTeam = (id: ITeam['id']) => action(types.REMOVE_TEAM, id);
+export const toggleTeamDisabled = (teamId: ITeam['id']) => action(types.TOGGLE_TEAM_DISABLED, teamId);
 export const recordTournamentResult = (data: TeamTournamentData, tournamentId: ITournament['id'], isForFame: boolean) =>
     action(types.RECORD_TOURNAMENT_RESULT, { data, tournamentId, isForFame });